Adding matches + clean up

+++ assets/js/classes/User.js
@@ -51,7 +51,7 @@ class User {
      */
     async register(body) {
         try {
-            const registerdUser = await FYSCloud.API.queryDatabase(
+            const registeredUser = await FYSCloud.API.queryDatabase(
                 "INSERT INTO users (firstName, lastName, email, residence, tel, password, birthDate, genderID) VALUES(?,?,?,?,?,?,?,?)",
                 [body.userFirstName.toLowerCase(), body.userLastName.toLowerCase(), body.userEmail.toLowerCase(),
                  body.userResidence.toLowerCase(),
@@ -60,8 +60,8 @@ class User {
 
 
             // saves user in browser
-            localStorage.setItem('FYSAuthId', registerdUser.insertId);
-            window.location.reload();
+            localStorage.setItem('FYSAuthId', registeredUser.insertId);
+            return registeredUser.insertId;
         } catch (e) {
             console.log(e);
         }
@@ -145,12 +145,32 @@ class User {
      */
     async updateUserImage(userID) {
         try {
-
             // delete existing file
-            var deletedFile = await FYSCloud.API.deleteFile(userID + ".png");
+            var hasProfileImage = await FYSCloud.API.fileExists(userID + ".png");
+
+            // checks if there is a profile image for this profile and deletes it
+            if (hasProfileImage) {
+                var deletedFile = await FYSCloud.API.deleteFile(userID + ".png");
+            }
 
             var file = await FYSCloud.Utils.getDataUrl($("#userUpdateImageFile"));
             return await FYSCloud.API.uploadFile(userID + ".png", file.url);
+        } catch (e) {
+            console.log(e);
+        }
+    }
+
+    /**
+     * Updates user bio
+     * @param userID
+     * @param content
+     * @returns {Promise<*>}
+     */
+    async updateBio(userID, content) {
+        try {
+            return await FYSCloud.API.queryDatabase(
+                "UPDATE users SET bio=? WHERE userID = ?",
+                [content, userID]);
         } catch (e) {
             console.log(e);
         }
@@ -218,7 +238,6 @@ class User {
      * @returns {Promise<void>}
      */
     async deleteInterest(type, typeID, userID) {
-        console.log("delete interest loaded")
         try {
             switch (type) {
                 case "hobbies":
@@ -237,5 +256,17 @@ class User {
         }
     }
 
+
+    /**
+     * Returns a random userID for the "I'm feeling lucky button"
+     * @returns {Promise<*>}
+     */
+    async getRandomUser() {
+        try {
+            return await FYSCloud.API.queryDatabase("SELECT userID FROM users ORDER BY RAND() LIMIT 1");
+        } catch (e) {
+            console.log(e);
+        }
+    }
 
 }
